+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+xxx
+===================
+- Added memory overload warning log message.
+
 0.11.8 / 2019-02-05
 ===================
 - Fix the `keyValueStore.forEachKey()` method.